requirements.txt
gitpython==3.1.36
fastapi==0.95.2
uvicorn==0.22.0
pyyaml==6.0
python-multipart==0.0.6
Pygments==2.16.1
typing-extensions==4.7.1
